feat(app): add a second route if ngRoute is chosen
This adds a second 'about' route if ngRoute is enabled. It also changes
around the layout to make the view replace the inner content instead of
the entire page as real world apps are more like this. Now the header
and footer are part of the app and only things between them change.